Initial Task Manager refactor
This refactors the task manager into a subpackage with multiple files,
similar to how the Workflow Manager is currently structured. Previously,
the Task Manager code was in a single file, making it hard to manage.
